% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIdata.r
\name{getVols}
\alias{getVols}
\title{get all the IDs of monitored reservoirs from FUNCEME API}
\usage{
getVols(id = 9, date_lower = strftime(Sys.time() - 7 * 60 * 60 * 24,
  format = "\%Y-\%m-\%d"), N = 1)
}
\description{
get all the IDs of monitored reservoirs from FUNCEME API
}
