% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valbuhayra-data.r
\docType{data}
\name{hydro_regions}
\alias{hydro_regions}
\title{Hydrographic regions of Ceara}
\format{An object of class \code{"sf"}.}
\usage{
data(hydro_regions)
}
\description{
Hydrographic regions of Ceara
}
\keyword{datasets}
