% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIdata.r
\name{latlong2id}
\alias{latlong2id}
\title{Request reservoir id based on jrc_neb table. only works on uni VPN and with password.}
\usage{
latlong2id(lon, lat, pw, hostname)
}
\description{
Request reservoir id based on jrc_neb table. only works on uni VPN and with password.
}
