% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valbuhayra-data.r
\docType{data}
\name{p_gauges_saved}
\alias{p_gauges_saved}
\title{List of precipitation gauges monitored by FUNCEME}
\format{An object of class \code{"sf"}.}
\source{
\href{http://api.funceme.br/help}{FUNCEME API}
}
\usage{
data(p_gauges_saved)
}
\description{
The data was obtained from the \href{http://api.funceme.br/help}{FUNCEME API} and the gauges are being updated everytime a new gauge is installed and goes online.
}
\examples{
data(p_gauges_saved)
\donttest{plot(p_gauges_saved["codigo1"])}
}
\keyword{datasets}
