% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIdata.r
\name{requestPrecip}
\alias{requestPrecip}
\title{Request rainfall measurements from FUNCEME API.}
\usage{
requestPrecip(codigo, requestDate, returnN)
}
\arguments{
\item{codigo}{is the id of the station, use data(p_gauges_saved) to obtain a list of available stations with metadata.}

\item{requestDate}{a data of class `Date`}

\item{returnN}{(number of points that should be returned)}
}
\value{
`data.frame` with columns `datetime`, `value` and `codigo`
}
\description{
Request rainfall measurements from FUNCEME API.
}
\author{
JM Delgado, \email{martinsd@uni-potsdam.de}
}
