% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIdata.r
\name{requestVolumes}
\alias{requestVolumes}
\title{Request measured volumes in strategic reservoirs from FUNCEME API. id, requestDate and returnN (number of points that should be returned)}
\usage{
requestVolumes(id, requestDate, returnN)
}
\description{
Request measured volumes in strategic reservoirs from FUNCEME API. id, requestDate and returnN (number of points that should be returned)
}
