% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valbuhayra-data.r
\docType{data}
\name{reservoirs}
\alias{reservoirs}
\title{Monitored reservoirs in Ceará}
\format{An object of class \code{"sf"}.}
\source{
\href{http://api.funceme.br/help}{FUNCEME API}
}
\usage{
data(reservoirs)
}
\description{
The data was obtained from the \href{http://api.funceme.br/help}{FUNCEME API} and the IDs were matched with \code{wm}. Column \code{cod} is the ID in the API and column \code{id} is the ID given in the watermask \code{wm}.
}
\examples{
data(reservoirs)
\donttest{plot(reservoirs["capacity"])}
}
\keyword{datasets}
