% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIdata.r
\name{sexagesimal2decimal}
\alias{sexagesimal2decimal}
\title{Get decimal from base 60. Accepts string and returns decimal}
\usage{
sexagesimal2decimal(string)
}
\description{
Get decimal from base 60. Accepts string and returns decimal
}
