% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valbuhayra-data.r
\docType{data}
\name{wm}
\alias{wm}
\title{Reservoir water extents in Ceará}
\format{An object of class \code{"sf"}. Column \code{ingestion_time} is in UTC. Column \code{area} is in $m^2$.}
\source{
\href{http://scihub.copernicus.eu/}{Sentinel API}
}
\usage{
data(wm)
}
\description{
Watermasks from May, June, July and August 2018 in Ceará, Northeast Brazil. The data was derived mainly from Sentinel data from ESA
with the package \href{https://github.com/jmigueldelgado/buhayra}{buhayra}. Most processing runs in the
computation cluster of the University of Potsdam, Germany.
}
\examples{
data(wm)
\donttest{plot(wm)}
}
\keyword{datasets}
