% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valbuhayra-data.r
\docType{data}
\name{CAV}
\alias{CAV}
\title{Height-Area-Volume curves for monitored reservoirs in Ceará}
\format{A dataframe.}
\source{
\href{http://api.funceme.br/help}{FUNCEME API}
}
\usage{
data(CAV)
}
\description{
The data was obtained from the \href{http://api.funceme.br/help}{FUNCEME API}.
}
\examples{
data(CAV)
\donttest{head(CAV)}
}
\keyword{datasets}
