% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAPIdata.r
\name{requestGauges}
\alias{requestGauges}
\title{Request list of rainfall gauges from funceme API}
\usage{
requestGauges(requestDate = today(), Ndays = 2)
}
\description{
Request list of rainfall gauges from funceme API
}
