% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valbuhayra-data.r
\docType{data}
\name{reservoirs}
\alias{reservoirs}
\title{Monitored reservoirs in Ceará}
\format{An object of class \code{"sf"}.}
\source{
\href{http://api.funceme.br/help}{FUNCEME API}
}
\usage{
data(reservoirs)
}
\description{
The data was obtained from the \href{http://api.funceme.br/help}{FUNCEME API}.
}
\examples{
data(reservoirs)
\donttest{plot(reservoirs["capacidade"])}
}
\keyword{datasets}
